refactor: 도메인,Repository 테스트 분리
